+++ shiny/ui/_valuebox.py
@@ -334,19 +334,23 @@ def value_box(
         :func:`~shiny.ui.showcase_left_center`, :func:`~shiny.ui.showcase_top_right()`,
         or :func:`~shiny.ui.showcase_bottom()` functions. Use the options functions when
         you want to control the height or width of the showcase area.
-     theme
+    theme
         The name of a theme (e.g. `"primary"`, `"danger"`, `"purple"`, `"bg-green"`,
         `"text-red"`) for the value box, or a theme constructed with
-        :func:`~shiny.ui.value_box_theme`. The theme names provide a convenient way to
-        use your app's Bootstrap theme colors as the foreground or background colors of
-        the value box. For more control, you can create your own theme with
-        :func:`~shiny.ui.value_box_theme` where you can pass foreground and background
-        colors directly. Bootstrap supported color themes: `"blue"`, `"purple"`,
-        `"pink"`, `"red"`, `"orange"`, `"yellow"`, `"green"`, `"teal"`, and `"cyan"`.
-        These colors can be used with `bg-NAME`, `text-NAME`, and
-        `bg-gradient-NAME1-NAME2` to change the background, foreground, or use a
-        background gradient respectively. If a `theme` string does not start with
-        `text-` or `bg-`, it will be auto prefixed with `bg-`.
+        :func:`~shiny.ui.value_box_theme`.
+
+        The theme names provide a convenient way to use your app's Bootstrap theme
+        colors as the foreground or background colors of the value box. For more
+        control, you can create your own theme with :func:`~shiny.ui.value_box_theme`
+        where you can pass foreground and background colors directly.
+
+        Bootstrap supported color themes: `"blue"`, `"purple"`, `"pink"`, `"red"`,
+        `"orange"`, `"yellow"`, `"green"`, `"teal"`, and `"cyan"`. These colors can be
+        used with `bg-NAME`, `text-NAME`, and `bg-gradient-NAME1-NAME2` to change the
+        background, foreground, or use a background gradient respectively.
+
+        If a `theme` string does not start with `text-` or `bg-`, it will be auto
+        prefixed with `bg-`.
     full_screen
         If `True`, an icon will appear when hovering over the card body. Clicking the
         icon expands the card to fit viewport size.